refactor: emphasize ui file structure as an isolated part
Started trying to improve the structure to make sure the ui has its own root. Optimally, only the UI path can import Gtk and Gdk, or other modules from the ui dir.
That would allow us to ensure linting rules to prevent Gtk and Gdk getting loaded in the CLI or extension runtimes (but we're not there yet).

*  Moved "utils" only used by ui into ui/helpers

The main issue is ulauncher/main.py importing Gtk in order to print the Gtk version on boot, and also unconditionally importing the app even if it's not used (such as --version or --help).

The second biggest issue is that extension_controller.py uses get_icon_path which has to load gtk in order to get the theme dir.

There are also some "utils" like xinit and ewmh that I think should be moved.
