+++ ulauncher/modes/extensions/extension_controller.py
@@ -22,8 +22,8 @@ from ulauncher.modes.extensions.extension_manifest import (
 )
 from ulauncher.modes.extensions.extension_remote import ExtensionRemote
 from ulauncher.modes.extensions.extension_runtime import ExtensionRuntime
+from ulauncher.ui.get_icon_path import get_icon_path
 from ulauncher.utils.eventbus import EventBus
-from ulauncher.utils.get_icon_path import get_icon_path
 from ulauncher.utils.json_utils import json_load
