+++ ulauncher/ui/get_icon_path.py
