+++ ulauncher/ui/helpers/hotkey_controller.py
