+++ ulauncher/ui/helpers/text_highlighter.py
