+++ ulauncher/ui/helpers/tray_icon.py
