+++ ulauncher/ui/load_icon_surface.py
@@ -22,7 +22,7 @@ def load_icon_surface(icon: str, size: int, scaling_factor: int = 1) -> ImageSur
     real_size = size * scaling_factor
     try:
         if not icon.startswith("/"):
-            from ulauncher.utils.get_icon_path import get_icon_path
+            from ulauncher.ui.get_icon_path import get_icon_path
 
             icon = get_icon_path(icon, real_size) or DEFAULT_EXE_ICON
         pixbuf = GdkPixbuf.Pixbuf.new_from_file_at_size(icon, real_size, real_size)
