+++ ulauncher/ui/preferences/views/about.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 from ulauncher import api_version, paths, version
 from ulauncher.gi import Gtk
+from ulauncher.ui.load_icon_surface import load_icon_surface
 from ulauncher.ui.preferences.views import BaseView, styled
 from ulauncher.utils.launch_detached import open_detached
-from ulauncher.utils.load_icon_surface import load_icon_surface
 
 ABOUT_VIEW_LOGO_SIZE = 200
