+++ ulauncher/ui/preferences/views/shortcuts.py
@@ -5,6 +5,7 @@ from time import time
 
 from ulauncher.gi import Gtk, Pango
 from ulauncher.modes.shortcuts.shortcuts import Shortcut, Shortcuts
+from ulauncher.ui.load_icon_surface import load_icon_surface
 from ulauncher.ui.preferences import views
 from ulauncher.ui.preferences.utils.ext_utils import autofmt_pango_code_block
 from ulauncher.ui.preferences.utils.sidebar_layout import SidebarItem, SidebarLayout
@@ -14,7 +15,6 @@ from ulauncher.ui.preferences.views import (
     get_window_for_widget,
     styled,
 )
-from ulauncher.utils.load_icon_surface import load_icon_surface
 
 
 class ShortcutsView(views.BaseView):
